#108 Update and fix tests
The problem was that the SchedulerWatcher which waits until the kill of the scheduler was notify before the scheduler which just verify where the scheduler call watcher. In that way, the scheduler does not have the time to update the second watcher that the assert verifications of the second watcher has been done and create a wrong test. TO fix this bug, we change the order of how watchers are added in scheduler. Now the watcher which verifies where the scheduler is pass is add before the watcher which wait the end of the scheduler, in that way we are sure that when the scheduler update the waiting watcher, the previous watcher has been also update before it.