#1f91vuf[review] Add the method MessageTransportProtocol#transport
Update MessageTransportProtocol -> now it is more clear that it is a protocol which manage the transport of message via the transport method.

Update SimpleBroadcastProtocol base on changes. Now the SimpleBroadcastProtocol use the transport method when the method broadcast is called.

Create TransportableIntendedToProtocol which is a mix between IntendedToProtocol and TransportableInEvent. This interface provides the fact the TransportableInEvent is intended to a protocol.

Update unit test

Update pom.xml

Signed-off-by: GuiGuiRakoto <guil.rako@hotmail.fr>