#1fen1qa[review] Review Event and MessageTransportProtocol
Now there is just Event. An Event can contain other Event

Message are now Event and can only contain Message (because a Message is an Event, it is not a problem)

MessageTransportProtocol is a protocol which transport message via the method MessageTransportProtocol#send. The class SimpleMessageSenderProtocol implement MessageTransportProtocol and simply send and receive message with no specific treatment. No guarantee on reception and order of delivery.

Update SimpleBroadcastProtocol which is now a simple Protocol which use a MessageTransportProtocol to broadcast messages.

Update unit tests

Update pom.xml version

Signed-off-by: GuiGuiRakoto <guil.rako@hotmail.fr>